+++ UI/window-basic-settings.hpp
@@ -37,6 +37,7 @@ class QComboBox;
 class QCheckBox;
 class QLabel;
 class QButtonGroup;
+class QSlider;
 class OBSPropertiesView;
 class OBSHotkeyWidget;
 
@@ -159,6 +160,10 @@ private:
 	QPointer<QRadioButton> vodTrack[MAX_AUDIO_MIXES];
 
 	QIcon hotkeyConflictIcon;
+	QCheckBox *statusOverlayEnabled = nullptr;
+	QComboBox *statusOverlayPosition = nullptr;
+	QSlider *statusOverlayOpacity = nullptr;
+	QLabel *statusOverlayOpacityValue = nullptr;
 
 	void SaveCombo(QComboBox *widget, const char *section, const char *value);
 	void SaveComboData(QComboBox *widget, const char *section, const char *value);
@@ -230,6 +235,8 @@ private:
 	OBSPropertiesView *CreateEncoderPropertyView(const char *encoder, const char *path, bool changed = false);
 
 	/* general */
+	void InitStatusOverlaySettingsUi();
+	void SetStatusOverlayControlsEnabled(bool enabled);
 	void LoadLanguageList();
 	void LoadThemeList(bool firstLoad);
 	void LoadBranchesList();